# Replacing math/rand with crypto/rand