Addresses:  https://github.com/stripe/goforit/issues/8

Before (with `math/rand`):
**Run1:**
```
goos: darwin
goarch: amd64
pkg: github.com/stripe/goforit
BenchmarkEnabled50-4    	 5000000	       263 ns/op
BenchmarkEnabled100-4   	20000000	        96.7 ns/op
PASS
ok  	github.com/stripe/goforit	3.846s
```
**Run2:**
```
goos: darwin
goarch: amd64
pkg: github.com/stripe/goforit
BenchmarkEnabled50-4    	 5000000	       273 ns/op
BenchmarkEnabled100-4   	20000000	        97.6 ns/op
PASS
ok  	github.com/stripe/goforit	3.930s
```

After (with `crypto/rand`):
**Run1:**
```
goos: darwin
goarch: amd64
pkg: github.com/stripe/goforit
BenchmarkEnabled50-4    	 1000000	      1461 ns/op
BenchmarkEnabled100-4   	20000000	        89.6 ns/op
PASS
ok  	github.com/stripe/goforit	3.765s
```
**Run2**
```
goos: darwin
goarch: amd64
pkg: github.com/stripe/goforit
BenchmarkEnabled50-4    	 1000000	      1444 ns/op
BenchmarkEnabled100-4   	20000000	        86.4 ns/op
PASS
ok  	github.com/stripe/goforit	3.607s
```