@bobby-stripe 2022-04-11
Sorry @pulkitsharma07 for letting this wilt on the vine!  As your benchmarks show, this change would Gets 5x slower.  In #43 we switched to using `sync.Pool` of math.Rand objects, seeded from `crypto/rand`, improving on the performance of using a single `*rand.Rand` with much much harder to predict sequences, so I'm going to close this for now.